requirements.txt
# Django Framework
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0  # Recommended for API auth

# Database (MySQL)
PyMySQL>=1.0.0
django-mysql>=4.0.0

# Security & Auth
django-csp>=4.0
django-axes>=8.0.0
django-cors-headers>=4.3.0  # Recommended for API

# Environment & Configuration
python-decouple>=3.8.0
django-environ>=0.12.0

# Machine Learning & Deep Learning
# Note: You already installed torch with CUDA 11.8
# These lines will ensure compatibility
torch>=2.7.1
torchvision>=0.22.1
torchaudio>=2.7.1
transformers>=4.40.0  # Updated version
accelerate>=0.20.0  # Required for efficient model loading and device_map
numpy<2.0.0,>=1.24.0  # Add upper bound for compatibility
pandas>=2.0.0

# NLP
spacy>=3.8.11
rapidfuzz>=3.0.0

# Computer Vision & OCR
easyocr>=1.7.0
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.24.0  # Often needed for image processing

# Document Processing
openai>=1.40.0
PyMuPDF>=1.23.0
pdf2image>=1.16.0
python-magic-bin>=0.4.14  # Windows compatible version
poppler-utils  # Required for pdf2image on Windows

# Utilities
pytz>=2023.3
python-dateutil>=2.8.0
tqdm>=4.66.0  # Progress bars for ML tasks

# Development
black>=24.0.0  # Code formatting
flake8>=7.0.0  # Linting